+++ bin/server.js
@@ -1,14 +1,16 @@
-const db = require("../db/db");
+const mongoose = require("mongoose");
+// const db = require("../db/db");
 require("dotenv").config();
 
 const app = require("../app");
 
 const { DB_HOST, PORT = 4000 } = process.env;
 
-db.connect(DB_HOST, {
-  useNewUrlParser: true,
-  useUnifiedTopology: true,
-})
+mongoose
+  .connect(DB_HOST, {
+    useNewUrlParser: true,
+    useUnifiedTopology: true,
+  })
   .then(() => {
     app.listen(PORT);
     console.log(`Server running. Use our API on port: ${PORT}`);
