+++ controllers/auth/signup.js
@@ -1,0 +1,36 @@
+const { Users } = require("../../repositories/users");
+const HttpCode = require("../../helpers/constants");
+require("dotenv").config();
+
+const signup = async (req, res, next) => {
+  const { email } = req.body;
+  const user = await Users.findByEmail(email);
+  if (user) {
+    return res.status(HttpCode.CONFLICT).json({
+      status: "error",
+      contentType: "application/json",
+      code: HttpCode.CONFLICT,
+      responseBody: {
+        message: "Email in use",
+      },
+    });
+  }
+  try {
+    const newUser = await Users.create(req.body);
+    return res.json({
+      status: "created",
+      contentType: "application/json",
+      code: HttpCode.CREATED,
+      responseBody: {
+        user: {
+          email: newUser.email,
+          subscription: newUser.subscription,
+        },
+      },
+    });
+  } catch (error) {
+    next(error);
+  }
+};
+
+module.exports = signup;
