+++ middlewares/index.js
@@ -1,7 +1,9 @@
 const controllerWrapper = require("./controllerWrapper");
 const joiValidation = require("./joiValidation");
+const authenticate = require("./authenticate");
 
 module.exports = {
   controllerWrapper,
   joiValidation,
+  authenticate,
 };
