--- validation/ctrl.js
@@ -1,11 +1,0 @@
-const ctrl = (ctrl) => {
-  return async (req, res, next) => {
-    try {
-      await ctrl(req, res, next);
-    } catch (error) {
-      next(error);
-    }
-  };
-};
-
-module.exports = ctrl;
