--- validation/index.js
@@ -1,5 +1,0 @@
-const validation = require("./validation");
-const ctrl = require("./ctrl.js");
-const authenticate = require("./authenticate");
-
-module.exports = { validation, ctrl, authenticate };
